% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComBat2.R
\name{ComBat2}
\alias{ComBat2}
\title{Adjust for batch effects using an empirical Bayes framework}
\usage{
ComBat2(dat, batch, mod = NULL, par.prior = FALSE, mean.only = FALSE,
  prior.plots = FALSE, maxGB = 2, log2transform = FALSE,
  transposed = FALSE, ref.batch = NULL)
}
\arguments{
\item{dat}{Genomic measure matrix (dimensions probe x sample). Should be class type bigmemory::big.matrix}

\item{batch}{{Batch covariate (only one batch allowed)}}

\item{mod}{Model matrix for outcome of interest and other covariates besides batch}

\item{par.prior}{(Optional) TRUE indicates parametric adjustments will be used, FALSE indicates non-parametric adjustments will be used}

\item{mean.only}{(Optional) FALSE If TRUE ComBat only corrects the mean of the batch effect (no scale adjustment)}

\item{prior.plots}{(Optional) TRUE give prior plots with black as a kernel estimate of the empirical batch effect density and red as the parametric}

\item{maxGB}{Max GB to use (batched up operations)}

\item{log2transform}{whether log2 transformation of dat is needed}

\item{transposed}{whether the dat matrix is transposed (samples x probes)}

\item{ref.batch}{(Optional) NULL If given, will use the selected batch as a reference for batch adjustment.}
}
\value{
data A probe x sample genomic measure matrix, adjusted for batch effects.
}
\description{
ComBat allows users to adjust for batch effects in datasets where the batch covariate is known, using methodology
described in Johnson et al. 2007. It uses either parametric or non-parametric empirical Bayes frameworks for adjusting data for
batch effects.  Users are returned an expression matrix that has been corrected for batch effects. The input
data are assumed to be cleaned and normalized before batch effect removal.
}
