% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_col_sum}
\alias{norm_col_sum}
\title{Column sum normalization of rectangular array}
\usage{
norm_col_sum(X)
}
\arguments{
\item{X}{A rectangular array.}
}
\value{
Normalized array with columns that sum to 1.
}
\description{
Normalize a rectangular array by dividing each element by the sum of its column.
This ensures that each column sum is 1.
}
