% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_row_sum}
\alias{norm_row_sum}
\title{Row sum normalization of rectangular array}
\usage{
norm_row_sum(X)
}
\arguments{
\item{X}{A rectangular array.}
}
\value{
Normalized array with rows that sum to 1.
}
\description{
Normalize a rectangular array by dividing each element by the sum of its row.
This ensures that each row sum is 1.
}
