% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_helpfunctions.R
\name{sesame_bisConversionControl}
\alias{sesame_bisConversionControl}
\title{Compute internal bisulfite conversion control
Compute GCT score for internal bisulfite conversion control. 
The higher the GCT score, the more likely the incomplete conversion.}
\usage{
sesame_bisConversionControl(meth, use.median = FALSE)
}
\description{
Compute internal bisulfite conversion control
Compute GCT score for internal bisulfite conversion control. 
The higher the GCT score, the more likely the incomplete conversion.
}
